chore: group validation with type
