+++ libs/entities/src/types/index.ts
@@ -7,5 +7,5 @@ export * from './place';
 export * from './product';
 export * from './purchase';
 export * from './supplier';
-export * from './item';
-export * from './recipe';
+export * from './auth';
+export * from './unit';
