+++ libs/entities/src/types/me/index.ts
@@ -1,2 +1,3 @@
 export * from './role.type';
 export * from './user.type';
+export * from './permission.type';
