+++ libs/entities/src/types/me/permission.type.ts
