--- libs/entities/src/types/recipe/index.ts
@@ -1,1 +1,0 @@
-export * from './recipe.type';
