+++ libs/entities/src/types/unit/index.ts
@@ -1,2 +1,1 @@
 export * from './unit-type.type';
-export * from './item.type';
