--- libs/entities/src/validation-schemas/index.ts
@@ -1,8 +1,0 @@
-export * from './item';
-export * from './recipe';
-export * from './common';
-export * from './order';
-export * from './me/permission.schema';
-export * from './me/role.schema';
-export * from './auth/login.schema';
-export * from './auth/logout.schema';
