--- libs/entities/src/validation-schemas/item/index.ts
@@ -1,2 +1,0 @@
-export * from './unit-type.schema';
-export * from './item.schema';
