--- libs/entities/src/validation-schemas/me/role.schema.ts
@@ -1,8 +1,0 @@
-import { z } from 'zod';
-
-export const roleResponseSchema = z.object({
-  id: z.string(),
-  name: z.string(),
-  createdAt: z.date().nullable(),
-  updatedAt: z.date().nullable(),
-});
