--- libs/entities/src/validation-schemas/order/index.ts
@@ -1,1 +1,0 @@
-export * from './order';
