--- libs/entities/src/validation-schemas/recipe/index.ts
@@ -1,2 +1,0 @@
-export * from './recipe.schema';
-export * from './recipe-detail.schema';
